refactor: move createCars to RacingGameController
