+++ src/main/java/controller/RacingGameController.java
@@ -7,12 +7,16 @@ import strategy.RandomMovableStrategy;
 import ui.InputView;
 import ui.ResultView;
 
+import java.util.*;
+
 public class RacingGameController {
+    private static final String NAME_DELIMITER = ",";
 
     public static void main(String[] args) {
-        Cars cars = new Cars(InputView.getCars());
+        String names = InputView.getNames();
         int tryCount = InputView.getTryCount();
 
+        Cars cars = new Cars(createCars(names));
         RandomMovableStrategy randomMovableStrategy = new RandomMovableStrategy();
         RacingGame racingGame = new RacingGame(randomMovableStrategy);
         ResultView.printStartMessage();
@@ -22,5 +26,13 @@ public class RacingGameController {
         }
 
         ResultView.printWinners(cars.getWinners());
+    }
+
+    static List<Car> createCars(String names) {
+        List<Car> cars = new ArrayList<>();
+        for (String name : Arrays.asList(names.split(NAME_DELIMITER))) {
+            cars.add(new Car(name));
+        }
+        return Collections.unmodifiableList(cars);
     }
 }
