+++ src/test/java/controller/RacingGameControllerTest.java
@@ -1,4 +1,4 @@
-package ui;
+package controller;
 
 import domain.Car;
 import org.junit.jupiter.api.Test;
@@ -7,11 +7,11 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class InputViewTest {
+public class RacingGameControllerTest {
     @Test
     public void createCarsTest() {
         String names = "a,b,c";
-        List<Car> cars = InputView.createCars(names);
+        List<Car> cars = RacingGameController.createCars(names);
         assertThat(cars).hasSize(3);
         assertThat(cars).contains(new Car("a"), new Car("b"), new Car("c"));
     }
